organize and moved some functions
